feat: Add functionality to restore last session on blank page
Added a condition to check if the page is blank and if a date is stored in
local storage. If the page is blank and a date is found, the user is redirected
to the saved date. If no date is found, an alert is shown prompting the user to
enter a date in the address bar.